PIF - 024 - Before Christ dates
I have added a checkbox near the date picker box to select if the date is before or after christ.

According to ISO - 8601 a date (before christ) must be rapresented with a minus before the date.

I have uploaded a book in phaidradev.cab.unipd.it with PID: o:8952 with the date format:

-YYYY-MM-DD

and the upload was successful.